Update KMeans init docstring to include 'random' option (#1020)
The docstring for the dask_ml.cluster.KMeans class currently omits 'random'
as a valid option for the init parameter. However, this option is fully
implemented in the underlying k_init function and serves as a critical
scalable alternative to the default 'k-means||', which can overwhelm the
scheduler on large datasets.